+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Pitch variation of sound effects.
+
 ### Changed
 - Switched to a more powerful audio driver: multiple sounds effects can
   now be played at once.